move: side-bar/index, SideBar.tsx, menu-item-iconsをside-bar→componentsに移動しました
