+++ components/menu-item-icons/HomeIcon.tsx
@@ -1,4 +1,4 @@
-import { menuItemIconProps } from "../../../interfaces";
+import { menuItemIconProps } from "../../interfaces";
 
 export const HomeIcon = ({ item, chosenItem }: menuItemIconProps) => {
   const isChose = item === chosenItem;
