+++ components/menu-item-icons/MenuIcon.tsx
@@ -1,4 +1,4 @@
-import { menuItemIconProps } from "../../../interfaces";
+import { menuItemIconProps } from "../../interfaces";
 
 export const MenuIcon = ({ item, chosenItem }: menuItemIconProps) => {
   const isChose = item === chosenItem;
