+++ components/menu-item-icons/MonitoringIcon.tsx
@@ -1,4 +1,4 @@
-import { menuItemIconProps } from "../../../interfaces";
+import { menuItemIconProps } from "../../interfaces";
 
 export const MonitoringIcon = ({ item, chosenItem }: menuItemIconProps) => {
   const isChose = item === chosenItem;
