+++ components/menu-item-icons/SettingsIcon.tsx
@@ -1,5 +1,5 @@
 import { url } from "inspector";
-import { menuItemIconProps } from "../../../interfaces";
+import { menuItemIconProps } from "../../interfaces";
 
 export const SettingsIcon = ({ item, chosenItem }: menuItemIconProps) => {
   const isChose = item === chosenItem;
