+++ components/menu-item-icons/ShiftIcon.tsx
@@ -1,4 +1,4 @@
-import { menuItemIconProps } from "../../../interfaces";
+import { menuItemIconProps } from "../../interfaces";
 
 export const ShiftIcon = ({ item, chosenItem }: menuItemIconProps) => {
   const isChose = item === chosenItem;
