+++ components/menu-item-icons/index.ts
