+++ pages/index.tsx
@@ -1,7 +1,7 @@
 import type { NextPage } from "next";
 import Head from "next/head";
 
-import { SideBar } from "../components/side-bar";
+import { SideBar } from "../components";
 
 const Home: NextPage = () => {
   return (
